requirements.txt
openai
tweepy
requests
beautifulsoup4
python-dotenv
